Refactored user and group models and controllers into separate files
