+++ server/groups/groupController.js
@@ -6,7 +6,10 @@ module.exports = {
   },
 
   create: function (req, res) {
-
+    Groups.sync().then(function () {
+      var group = Groups.build(req.body);
+      group.save();
+    });
   },
 
   browse: function (req, res) {
@@ -14,7 +17,13 @@ module.exports = {
   },
 
   join: function (req, res) {
-
+    addToGroup = function(groupName, username){
+      Groups.findOne({where: {name: groupName}}).then(function (group) {
+        Users.findOne({where: {username: username}}).then(function (user) {
+          user.addGroups(group);
+        });
+      });
+    };
   },
 
   ping: function (req, res) {
