+++ server/users/userModel.js
@@ -1,62 +1,12 @@
-var mongoose = require('mongoose'),
-    bcrypt   = require('bcrypt'),
-    Q        = require('q'),
-    SALT_WORK_FACTOR  = 10;
+var sequelize = require('./db/index.js');
+var Groups = require('./groups/groupModel.js');
 
-
-var UserSchema = new mongoose.Schema({
-  username: {
-    type: String,
-    required: true,
-    unique: true
-  },
-
-  password: {
-    type: String,
-    required: true
-  },
-  salt: String
+var Users = sequelize.define('Users', {
+  username: {type: Sequelize.STRING, unique: true} //make unique
+  email: Sequelize.STRING,
+  phone: Sequelize.BIGINT
 });
 
-UserSchema.methods.comparePasswords = function (candidatePassword) {
-  var defer = Q.defer();
-  var savedPassword = this.password;
-  bcrypt.compare(candidatePassword, savedPassword, function (err, isMatch) {
-    if (err) {
-      defer.reject(err);
-    } else {
-      defer.resolve(isMatch);
-    }
-  });
-  return defer.promise;
-};
+Users.belongsToMany(Groups, {through: 'Memberships'});
 
-UserSchema.pre('save', function (next) {
-  var user = this;
-
-  // only hash the password if it has been modified (or is new)
-  if (!user.isModified('password')) {
-    return next();
-  }
-
-  // generate a salt
-  bcrypt.genSalt(SALT_WORK_FACTOR, function(err, salt) {
-    if (err) {
-      return next(err);
-    }
-
-    // hash the password along with our new salt
-    bcrypt.hash(user.password, salt, function(err, hash) {
-      if (err) {
-        return next(err);
-      }
-
-      // override the cleartext password with the hashed one
-      user.password = hash;
-      user.salt = salt;
-      next();
-    });
-  });
-});
-
-module.exports = mongoose.model('users', UserSchema);
+exports.Users = Users;
